Common storage/messagebroker fixture was moved to the connectors package and splitted into 2 separate setups.
